+++ R/futurize.R
@@ -103,7 +103,9 @@ futurize <- function(expr, substitute = TRUE, options = futurize_options(...), .
     on.exit(mdebug_pop())
   }
 
-  transpile(expr, substitute = FALSE, options = options, when = when, eval = eval, type = "futurize::add-on", envir = envir, what = "futurize", debug = debug)
+  disable <- !isTRUE(getOption("futurize.enable", TRUE))
+  
+  transpile(expr, substitute = FALSE, options = options, when = when, eval = eval, disable = disable, type = "futurize::add-on", envir = envir, what = "futurize", debug = debug)
 } ## futurize()
 class(futurize) <- c("transpiler", class(futurize))